// src/generated/models/microsoft-entra-id-model.ts
export interface MicrosoftEntraID {
  /**
   * @displayName Id.
   * @description Unique system identifier - GUID generated by system
   * @uiConstraints NEVER create input fields for this ID
   */
  id: string;
  /**
   * @displayName Nombre para mostrar
   * @description Human-readable record identifier
   */
  nombreparamostrar: string;
  /**
   * @displayName Apellido
   */
  apellido?: string;
  /**
   * @displayName Código postal
   */
  cdigopostal?: string;
  /**
   * @displayName Ciudad
   */
  ciudad?: string;
  /**
   * @displayName Correo
   */
  correo?: string;
  /**
   * @displayName Cuenta de Microsoft Entra ID habilitada
   */
  cuentademicrosoftentraidhabilitada?: boolean;
  /**
   * @displayName Dirección de la calle
   */
  direccindelacalle?: string;
  /**
   * @displayName Fecha y hora de creación
   */
  fechayhoradecreacin?: string;
  /**
   * @displayName Idioma preferido
   */
  idiomapreferido?: string;
  /**
   * @displayName ImAddresses
   */
  imaddresses?: string;
  /**
   * @displayName Nombre de empresa
   */
  nombredeempresa?: string;
  /**
   * @displayName Nombre de pila
   */
  nombredepila?: string;
  /**
   * @displayName Nombre principal de usuario
   */
  nombreprincipaldeusuario?: string;
  /**
   * @displayName Puesto
   */
  puesto?: string;
  /**
   * @displayName Teléfono móvil
   */
  telfonomvil?: string;
  /**
   * @displayName Teléfonos del trabajo
   */
  telfonosdeltrabajo?: string;
  /**
   * @displayName Tipo de usuario
   */
  tipodeusuario?: string;
  /**
   * @displayName Ubicación de la oficina
   */
  ubicacindelaoficina?: string;
  /**
   * @displayName Un identificador único para Microsoft Entra ID
   * @validationRule Required for create/update operations
   */
  unidentificadornicoparamicrosoftentraid: string;
}

export const _MicrosoftEntraID = 'MicrosoftEntraID' as const;